Restructure the app
